requirements.txt
customtkinter>=5.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
pyyaml>=6.0
google-genai>=1.0.0
jinja2>=3.1.0
pyinstaller>=6.0.0
